some racey fixes

+++ osquery/runtime.go
@@ -8,6 +8,7 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/kolide/launcher/osquery/table"
@@ -301,7 +302,10 @@ func (o *OsqueryInstance) Recover() error {
 	// First, we try to kill the osqueryd process if it isn't already dead.
 	if !o.cmd.ProcessState.Exited() {
 		if err := o.cmd.Process.Kill(); err != nil {
-			return errors.Wrap(err, "could not kill the osquery process during recovery")
+			if !strings.Contains(err.Error(), "process already finished") {
+				return errors.Wrap(err, "could not kill the osquery process during recovery")
+
+			}
 		}
 	}
